+++ convex/convex.config.ts
@@ -1,13 +1,12 @@
 import resend from "@convex-dev/resend/convex.config";
 import staticHosting from "@convex-dev/static-hosting/convex.config";
 import auth from "@estifanos-sh/convex-auth/convex.config";
-import { authEnv } from "@estifanos-sh/convex-auth/server";
 import { defineApp } from "convex/server";
 import { v } from "convex/values";
 
 const app = defineApp({
   env: {
-    ...authEnv,
+    APP_URL: v.optional(v.string()),
     GOOGLE_CLIENT_ID: v.optional(v.string()),
     GOOGLE_CLIENT_SECRET: v.optional(v.string()),
     RESEND_API_KEY: v.string(),